move argparsing tests to own module
